refactor(payments): polish
