+++ packages/fxa-payments-server/src/store/defaultState.ts
@@ -12,7 +12,7 @@ import {
   UpdateSubscriptionPlanResult,
 } from './types';
 
-export default {
+export const defaultState = {
   customer: uninitializedFetch<Customer, APIError>(),
   plans: uninitializedFetch<Array<Plan>, APIError>(),
   profile: uninitializedFetch<Profile, APIError>(),
@@ -31,6 +31,9 @@ export default {
   updatePayment: uninitializedFetch<any>(),
 };
 
+// https://blog.usejournal.com/writing-better-reducers-with-react-and-typescript-3-4-30697b926ada
+export type State = typeof defaultState;
+
 // Utility to construct an uninitialized FetchState with type annotations
 // for future fetches
 function uninitializedFetch<T = any, E = any>(): FetchState<T, E> {
@@ -38,5 +41,7 @@ function uninitializedFetch<T = any, E = any>(): FetchState<T, E> {
     error: null,
     loading: false,
     result: null,
-  };
+  } as const;
 }
+
+export default defaultState;
